Move backup info extraction function to backupinfo file
